Port to machine agents shipping Java 11
Java 11 does not supply a SOAP implementation, unlike Java 8. A choice
was made to go with the latest libraries from the Eclipse Metro project
to avoid security vulnerabilities associated with old and deprecated
versions. This breaks compatibility with Java 8 and Java 11+ is now
required, which explains bumping the major version of the extension.

This change also deals with issues resulting from the way the machine
agent loads extensions, along with interactions between thread pools and
class loaders in Java 9 and above. It ensures SOAP handling
initialization does not happen in the constructor of the extension's
main class, otherwise the class loader will fail to locate certain
classes when launched via the machine agent (although tests pass).